add type declaration alongside files
