+++ package.json
@@ -4,15 +4,16 @@
   "description": "Interfaces for JS Libp2p",
   "leadMaintainer": "Jacob Heun <jacobheun@gmail.com>",
   "main": "src/index.js",
-  "types": "index.d.ts",
+  "types": "types/src/index.d.ts",
   "files": [
     "src",
-    "dist",
-    "index.d.ts"
+    "types",
+    "dist"
   ],
   "scripts": {
     "lint": "aegir lint",
     "build": "aegir build",
+    "pregenerate:types": "rimraf './src/**/*.d.ts'",
     "generate:types": "tsc",
     "test": "aegir test",
     "test:node": "aegir test --target node",
@@ -68,6 +69,7 @@
   "devDependencies": {
     "aegir": "^25.0.0",
     "it-handshake": "^1.0.1",
+    "rimraf": "^3.0.2",
     "typescript": "3.7.5"
   },
   "contributors": [
