+++ src/connection/index.d.ts
@@ -1,0 +1,1 @@
+export var Connection: typeof import('./connection');
