+++ src/connection/status.d.ts
@@ -1,0 +1,3 @@
+export declare const OPEN: string;
+export declare const CLOSING: string;
+export declare const CLOSED: string;
