+++ src/crypto/errors.d.ts
@@ -1,0 +1,15 @@
+export class UnexpectedPeerError extends Error {
+    static get code(): string;
+    constructor(message?: string);
+    code: string;
+}
+export class InvalidCryptoExchangeError extends Error {
+    static get code(): string;
+    constructor(message?: string);
+    code: string;
+}
+export class InvalidCryptoTransmissionError extends Error {
+    static get code(): string;
+    constructor(message?: string);
+    code: string;
+}
