+++ src/pubsub/errors.d.ts
@@ -1,0 +1,4 @@
+export namespace codes {
+    export const ERR_MISSING_SIGNATURE: string;
+    export const ERR_INVALID_SIGNATURE: string;
+}
