+++ src/pubsub/message/index.d.ts
@@ -1,0 +1,5 @@
+export var rpc: any;
+export var td: any;
+export var RPC: any;
+export var Message: any;
+export var SubOpts: any;
