+++ src/pubsub/message/topic-descriptor.proto.d.ts
@@ -1,0 +1,2 @@
+declare const _exports: string;
+export = _exports;
