+++ src/transport/errors.d.ts
@@ -1,0 +1,6 @@
+export class AbortError extends Error {
+    static get code(): string;
+    static get type(): string;
+    code: string;
+    type: string;
+}
